# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "verson": "1.0.1",
    "title": "E-commerce API",
    "description": "API documentation for E-commerce",
    "version": "1.0.0"
  },
  "host": "localhost:3001/api",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter JWT or API Key with Bearer into field"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/send-newletter": {
      "post": {
        "tags": [
          "Telegram"
        ],
        "description": "API to send newletter to telegram",
        "parameters": [
          {
            "name": "newletter",
            "in": "body",
            "description": "Newletter content",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "phoneNumber": {
                  "type": "string",
                  "example": "0123456789"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/send-product-order": {
      "post": {
        "tags": [
          "Telegram"
        ],
        "description": "API to send san-pham order to telegram",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Product information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "product": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "product name"
                    },
                    "id": {
                      "type": "string",
                      "example": "san-pham id"
                    }
                  }
                },
                "variantChosen": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "variant name"
                    },
                    "id": {
                      "type": "string",
                      "example": "variant id"
                    }
                  }
                },
                "productPrice": {
                  "type": "string",
                  "example": "product price"
                },
                "info": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "email"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "first name"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "last name"
                    },
                    "phoneNumber": {
                      "type": "string",
                      "example": "phone number"
                    },
                    "address": {
                      "type": "string",
                      "example": "address"
                    },
                    "specificAddress": {
                      "type": "string",
                      "example": "specific address"
                    },
                    "district": {
                      "type": "string",
                      "example": "district"
                    },
                    "city": {
                      "type": "string",
                      "example": "city"
                    },
                    "postcode": {
                      "type": "string",
                      "example": "postcode"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}